Re-use existing forked thread when exists
When replying to or mentioning a null-thread message, re-use the thread
id of any existing forked thread from a previous reply to that message
if there is one, so that not everyone replying to a null thread message
ends up generating a new thread when most of the time they mean to all
end up in the same new thread.

Of course user may tap thread icon at any time if they actually want a
second fork.

Mentions still do not generate a new thread since their association is
not tracked, but will follow the forked thread if it exists.