Retry onboarding completion
If anything goes wrong, we just log.  If app got killed in the middle of
the process, we don't notice.  So now instead we retry after 3 seconds
whenever something goes wrong and we retry when the app comes back to
the home screen if it didn't happen yet.  Still careful to only ever do
it once.