Fix fitzpatrick double-show
For some reason, emoji with skintone modifier showed as doubled, a big
one and then a little one.  The span was being set correctly over the
whole thing, and yet this happened.  If we set it to one longer, it
fixes the problem, and in all testing so far doesn't cause any new
problems. I don't love this because it doesn't fully make sense to me,
but it works so I'm putting it in for now.